1. move allow_internal_unstable to rustc_attr 2. as a result, drop rustc_errors dep from syntax
